packages/opencloning-db/pyproject.toml
[tool.poetry]
authors = ["Manuel Lera-Ramirez <manulera14@gmail.com>"]
description = "Database for OpenCloning, a web application to generate molecular cloning strategies in json format, and share them with others."
license = "MIT"
name = "opencloning-db"
version = "0.1.0"
repository = "https://github.com/manulera/OpenCloning_db"

[project]
name = "opencloning-db"
version = "0.1.0"
description = "Database for OpenCloning, a web application to generate molecular cloning strategies in json format, and share them with others."
requires-python = ">=3.11,<4"
dependencies = [
    "opencloning",
    "sqlalchemy>=2.0.41,<3",
    "fastapi-pagination>=0.15.10,<0.16",
    "PyJWT>=2.10.1,<3",
    "pwdlib[argon2]>=0.2.1,<0.3",
    "pydna>=5.5.9",
    "email-validator>=2.3.0,<3",
]

[tool.uv.sources]
opencloning = { workspace = true }

[tool.poetry.dependencies]
python = "^3.11"
opencloning = {path = "../../opencloning", develop = true}
sqlalchemy = "^2.0.41"
fastapi-pagination = "^0.15.10"
PyJWT = "^2.10.1"
pwdlib = {version = "^0.2.1", extras = ["argon2"]}
pydna = "^5.5.9"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
flake8-bugbear = "^24.2.6"
black = "^24.2.0"
pre-commit = "^3.6.2"
watchfiles = "^0.21.0"
nbstripout = "^0.8.1"

[tool.poetry.group.test.dependencies]
pytest = "8.3.4"
pre-commit = "^3.6.2"
pytest-cov = "^4.1.0"
pytest-rerunfailures = "^14.0"
respx = "^0.21.1"


[tool.poetry.group.ipython.dependencies]
ipython = "^8.20.0"
ipykernel = "^6.28.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
skip-string-normalization = true
line-length = 119
